Bug squishy
When starting with a privot defined, and the camera in the wrong resultant position, the displayed result would remain incorrect until the first input event was received to properly define the _triggered boolean. By forcing an update after enabling the script before the first input event is required corrects this bug.